Preserve executable bit in files
`identify` uses the executable bit of a file when generating file tags,
so `executable` is only listed when that bit is set.
Some commit hooks such as
`pre-commit-hooks/check-executables-have-shebangs` are designed to run
only on executable files via `types = ["executable"]`.

By removing `--no-preserve=mode` we get to keep the executable bit. We
still need everything to be writable though.